requirements.txt
# Momotopsy
# Core API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Document Ingestion
PyMuPDF>=1.24.0
python-docx>=1.1.0
easyocr>=1.7.1

# ML / Math
sentence-transformers>=2.6.0
networkx>=3.2
scikit-learn>=1.4.0
numpy>=1.26.0
joblib>=1.3.0

# Training Pipeline
datasets>=2.18.0
pandas>=2.2.0
imbalanced-learn>=0.12.0